// src/common/types/tablageneral.type.ts
export interface TablageneralType {
  equipoId: number;
  equipo: string;
  logo: string;
  puntos: number;
  goles: number;
  golesRecibidos: number;
  partidosJugados: number;
  partidosGanados: number;
  partidosGanadosDesempate: number;
  partidosPerdidos: number;
  partidosPerdidosDesempate: number;
  partidosEmpatados: number;
}